HOLY SHIT COMMIT MESSAGES
Sounds and shit.
CoreScripts needs the SoundManager script, which then needs to have all the sounds added to it. It also needs the AudioSource prefab.

If the AudioSource prefab breaks, it's just a gameObject with the AudioPlayer script and an AudioSource component.

Also made it so you can't toggle between movement types if you're not actually able to move a platform.